requirements.txt
pillow
pytesseract
tensorflow
requests
fal-client

pytest
